+++ chrome/extension/background.js
@@ -31,4 +31,4 @@ promisifyAll(chrome.storage, [
 require('./background/contextMenus');
 require('./background/inject');
 require('./background/badge');
-require('./background/keyBindings');
+require('./background/app');
