+++ chrome/views/background.pug
@@ -2,4 +2,8 @@ doctype html
 
 html
   head
+    meta(charset='UTF-8')
+    title Heutagogy
+  body
+    #root
     script(src=env == 'prod' ? '/js/background.bundle.js' : 'http://localhost:3000/js/background.bundle.js')
