+++ webpack/dev.config.js
@@ -9,7 +9,7 @@ const hotScript = 'webpack-hot-middleware/client?path=__webpack_hmr&dynamicPubli
 const baseDevConfig = () => ({
   devtool: 'eval-cheap-module-source-map',
   entry: {
-    app: [customPath, hotScript, path.join(__dirname, '../chrome/extension/app')],
+    app: [customPath, hotScript, path.join(__dirname, '../chrome/extension/background/app')],
     background: [customPath, hotScript, path.join(__dirname, '../chrome/extension/background')],
   },
   devMiddleware: {
