+++ webpack/prod.config.js
@@ -5,7 +5,7 @@ const customPath = path.join(__dirname, './customPublicPath');
 
 module.exports = {
   entry: {
-    app: [customPath, path.join(__dirname, '../chrome/extension/app')],
+    app: [customPath, path.join(__dirname, '../chrome/extension/background/app')],
     background: [customPath, path.join(__dirname, '../chrome/extension/background')],
     inject: [customPath, path.join(__dirname, '../chrome/extension/inject')],
   },
